requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
jinja2
python-multipart

requests
openai
langchain
langchain-openai
langchain-core

fastapi
uvicorn
jinja2
sqlalchemy
python-multipart
pydantic>=2
httpx
openai>=1.40.0

